Removed all the junk that has been (apparently) generated at /cmake
Also made some changes to the cmake script to be able to compile on Windows: I was able to do it on linux before but on Windows it gave me an annoying command-line error. Later on it gave me some linking errors due to MinGW version conflicts and after attempting to get it done I've aborted it (MSYS2 sucks for that). Now I'm using WSL with the MinGW toolchain to compile for Windows